requirements.txt
langchain
openai
PyPDF2
faiss-cpu
tiktoken
streamlit
python-dotenv
pickle
os
